Made severe changes to main.cpp, Settings.h and Settings.cpp
1. Changed names of functions within Settings - e.g.,
	initializeSettings_Diversification()
	InitializeSettings_Traits

to convention w "*_Traits" and "*_Diversification" to indicate which functions
deal with diversification and trait evolution

2. Created a single general input function, initializeSettingsDevel
that handles all file input from control file. NOw this requires controlfile
to have an argument modeltype, which is either "trait" or "speciationextinction".
All the parnames and values from controfile are read into Settings private variables,
_varNames and _varValues.

3. At end of initializeSettingsDevel, it calls initializeSettings_Diversification
or initializeSettings_Traits depending the value of variable _modeltype, a new
private variable in class Settings that indicates whether we are running speciationextinction
or trait model.

4. main.cpp is now updated to handle these changes.