Major file changes: added autotune class, which may not be useful. Other changes
1. Changed a major component of class BranchEvent and TraitBranchEvent. Formerly, these
classes handled their own position updates for local moves. Class model called an eventLocalMove()
function, which caused the BranchEvent class to move its own position.
The problem with this is that the operator (_scale or epsilon) for the sliding window
shift move for the eventLocalMove was assigned to the event in the event constructor.
Thus, if you want to change the bounds of the sliding window proposal during (say) autotuning
this would not be possible, since events have a fixed operator parameter for local moves.
The new construction eliminates this aspect of class BranchEvent and TraitBranchEvent.
Sliding window moves are now managed in class MOdel and TraitModel. From model,
we now call BrancEvent->moveEventLocal(stepsize) where stepsize is the target step.

2. Autotune class added. Only works for speciationextinction at moment. If you specify
autotune = 1 in controlfile it will run a brief analysis that outputs an operators file
that can then be imported into R and analyzed with the new function (see below)

3. Added R function for analyzing autotune file. Only works for speciationextinction.
Not sure that this works well, but worth exploring.

4. Changed name of MCMC operator for the sliding window operation that moves the position
of an event. Previously, a parameter meanSpeciationLength was the user-defined parameter
that specified magnitude of this step. This parameter has changed names to
updateEventLocationScale, which is more descriptive.