Add Log class
Instead of writing to std::cout directly, use the log()
function, which is defined when you include "Log.h".
You do not need to create a Log class, just use log().

For example, to write a plain message to standard output:

  log() << "This is a message.\n";
  log(Message) << "This does the same as above.\n";

To write a warning or error to standard error:

  log(Warning) << "This is a warning.\n";
  log(Error) << "This is an error.\n";

You can specify the output stream to write to.
For example, to write to a file named "log.txt":

  std::ofstream out("log.txt");
  log(out) << "This is a message.\n";
  log(Message, out) << "This does the same as above.\n";
  log(Warning, out) << "This is a warning.\n";
  log(Error, out) << "This is an error.\n";
  out.close();