Documentation: add a welcome page
To create a welcome page, it was necessary to modify the sphinxdoc theme
and create a new one (bamm). Because the welcome page should have the
index.html name, the old index.rst was renamed to documentation.rst. The
new index.html is generated from a template index.html located in
_templates. A layout.html file was also created to put a top banner on
every page of the website. The sphinx doc repository was used as a
reference to discover where to put files and how to edit them.