Allow parameters to be changed at command-line
Now, any parameter from the control file can be overridden by a
command-line argument. For example, the seed can be changed at the
command-line like this:

    ./bamm -c divcontrol.txt --seed 1234

The command-line argument must start with "--", followed by the name
of the paramater (as it is in the control file), followed by the new
value.