Add data writer to output proposal acceptance info
This feature used to be present, but was lost during refactoring. User
must add the option "outputAcceptanceInfo = 1" in the control file in
order to output the acceptance info. The default file name is
"acceptance_info.txt", but can be changed with the option
"acceptanceInfoFileName = xxxx.xxx", where xxxx.xxx is the name of the
file.