Introduce class Random, a wrapper for MbRandom
There are random-related methods, such as trueWithProbability that I
wanted to add, but I didn't want to do it in MbRandom because it is not
a BAMM class; it was obtain from somewhere else.

Some methods related to probability density functions were moved to the
Stat class because it doesn't make sense to keep them in a random-
related class. No randomness is used in these methods.