Choose random internal node starting at 0
There is no reason to start at 1, unless the intent was for the root
to never be chosen. However, because the internalNodeSet is a set and
not a vector, starting at 1 doesn't do anything because nodes in a set
are ordered by their memory location, regardless in the order they're
added to the set.