Major fixes: numerical underflow in likelihood calculations
Likelihood calculations as implemented for fossil BAMM (which reduce to
non-fossil BAMM) allowed overflow issues to occur whenever E0 became
numerically equal to 1.0.

This situation was previously addressed by having a check for a maximum
possible value for the instantaneous E(t) value on branches. However,
the check failed whenever there were multiple events on a branch. Hence,
the problematic scenario (with E0 -> 1.0) was not caught, leading to incorrect
likelihood calculations.

Now:
SpExModel::computeSpExProbBranch tests for E0 > extinctionProbMax,
and also for spProb values > 1.0. These are automatically rejected
as they now return -INFINITY.

Model::isEventConfigurationValid was also failing to correctly check
for invalid triads of events on adjacent branches. This is fixed.

BAMM also now provides user the option of adding a set of events at the
start of a run where the parameters of each are set to the initial values
specified in the control file. The previous version sampled parameter values
from the prior whenever the initial number of events was > 0.